requirements.txt
absl-py==2.1.0
accelerate==1.7.0
aiofiles==0.7.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiohttp-retry==2.9.1
aioredis==2.0.0
aiosignal==1.3.2
aiosmtplib==1.1.6
aiosqlite==0.17.0
altair==5.5.0
annotated-types==0.7.0
anyio==3.3.1
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
arrow==1.3.0
asgiref==3.8.1
asttokens==3.0.0
astunparse==1.6.3
async-lru==2.0.5
async-timeout==4.0.3
attrs==24.2.0
babel==2.17.0
backoff==2.2.1
beautifulsoup4==4.13.4
bidict==0.23.1
bleach==6.2.0
blinker==1.8.2
branca==0.7.2
bs4==0.0.2
buildozer==1.5.0
cachetools==5.5.0
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
classes==0.4.1
click==8.1.7
cmake==3.30.2
colorama==0.4.6
coloredlogs==15.0.1
comm==0.2.2
comtypes==1.4.9
config==0.5.1
contourpy==1.2.1
cryptography==42.0.8
cycler==0.12.1
dash==3.0.4
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
dataclasses==0.6
DateTime==5.5
debugpy==1.8.14
decorator==5.2.1
deepmerge==2.0
defusedxml==0.7.1
Detector==2.0
diffusers==0.33.1
distlib==0.3.8
Django==5.0.7
dlib @ file:///C:/Users/KIRAN/OneDrive/Documents/dlib-19.24.99-cp312-cp312-win_amd64.whl#sha256=20c62e606ca4c9961305f7be3d03990380d3e6c17f8d27798996e97a73271862
dnspython==2.6.1
dotmap==1.3.30
email_validator==2.2.0
et-xmlfile==1.1.0
executing==2.2.0
face-recognition==1.3.0
face_recognition_models==0.3.0
fastapi==0.111.1
fastapi-cli==0.0.4
fastjsonschema==2.21.2
filelock==3.15.4
Flask==3.0.3
Flask-Cors==5.0.0
Flask-Mail==0.10.0
Flask-MySQLdb==2.0.0
Flask-SocketIO==5.5.0
flatbuffers==24.3.25
folium==0.17.0
fonttools==4.53.1
fqdn==1.5.1
frozenlist==1.6.0
fsspec==2024.6.1
future==1.0.0
gast==0.6.0
gitdb==4.0.11
GitPython==3.1.43
google-ai-generativelanguage==0.6.10
google-api-core==2.24.0
google-api-python-client==2.158.0
google-auth==2.37.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.3
google-pasta==0.2.0
googleapis-common-protos==1.66.0
greenlet==3.1.1
grpcio==1.69.0
grpcio-status==1.69.0
gTTS==2.5.4
gunicorn==23.0.0
h11==0.14.0
h5py==3.11.0
httpcore==1.0.5
httplib2==0.22.0
httptools==0.6.1
httpx==0.27.0
huggingface-hub==0.31.4
humanfriendly==10.0
idna==3.7
ImageIO==2.37.2
imageio-ffmpeg==0.6.0
importlib_metadata==8.7.0
imutils==0.5.4
intel-openmp==2021.4.0
ipykernel==6.29.5
ipython==9.2.0
ipython_pygments_lexers==1.1.1
ipywidgets==8.1.7
isoduration==20.11.0
itsdangerous==2.2.0
jax==0.4.34
jaxlib==0.4.34
jedi==0.19.2
Jinja2==3.1.4
joblib==1.4.2
json5==0.12.1
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter==1.1.1
jupyter-console==6.6.3
jupyter-events==0.12.0
jupyter-lsp==2.3.0
jupyter_client==8.6.3
jupyter_core==5.8.0
jupyter_server==2.17.0
jupyter_server_terminals==0.5.3
jupyterlab==4.4.7
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.15
jwt==1.3.1
keras==3.4.1
kiwisolver==1.4.5
libclang==18.1.1
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.1
matplotlib-inline==0.1.7
mdurl==0.1.2
mediapipe==0.10.14
mistune==3.1.4
mkl==2021.4.0
ml-dtypes==0.4.0
MouseInfo==0.1.3
moviepy==2.2.1
mpmath==1.3.0
multidict==6.4.3
mysql==0.0.3
mysql-connector==2.2.9
mysql-connector-python==9.1.0
mysqlclient==2.2.6
namex==0.0.8
narwhals==1.20.1
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.3
nltk==3.9.1
notebook==7.4.5
notebook_shim==0.2.4
numpy==1.26.4
nvidia-cuda-runtime-cu12==12.9.37
onnxruntime==1.23.2
opencage==2.4.0
opencv-contrib-python==4.10.0.84
opencv-contrib-python-headless==4.10.0.84
opencv-python==4.8.0.76
opencv-python-headless==4.11.0.86
openpyxl==3.1.5
opt-einsum==3.3.0
optree==0.12.1
packaging==24.1
pandas==2.2.2
pandocfilters==1.5.1
parso==0.8.4
passlib==1.7.4
pexpect==4.9.0
pgpdump==1.5
phonenumbers==8.13.40
pillow==11.3.0
platformdirs==4.2.2
playsound==1.3.0
plotly==6.1.1
proglog==0.1.12
prometheus_client==0.23.1
prompt_toolkit==3.0.51
propcache==0.3.1
proto-plus==1.25.0
protobuf==4.25.8
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
py-cpuinfo==9.0.0
pyarrow==18.1.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
PyAudio==0.2.14
PyAutoGUI==0.9.54
pycparser==2.22
pydantic==2.8.2
pydantic_core==2.20.1
pydeck==0.9.1
pyfingerprint==1.5
pygame==2.6.0
PyGetWindow==0.0.9
pyglet==2.0.16
Pygments==2.18.0
PyJWT==2.10.1
PyMsgBox==1.0.9
PyMySQL==1.1.1
pynput==1.7.7
pyparsing==3.1.2
pyperclip==1.9.0
pypiwin32==223
pyreadline3==3.4.1
PyRect==0.2.0
PyScreeze==1.0.1
pyserial==3.5
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-engineio==4.11.2
python-json-logger==3.3.0
python-multipart==0.0.9
python-socketio==5.12.1
pyttsx3==2.98
pytweening==1.2.0
pytz==2024.1
pywin32==308
pywinpty==3.0.0
PyYAML==6.0.1
pyzmq==26.4.0
redis==5.2.1
referencing==0.35.1
regex==2024.11.6
requests==2.32.3
retrying==1.3.4
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.7.1
rpds-py==0.22.3
rsa==4.9
safetensors==0.5.3
scikit-learn==1.5.1
scipy==1.14.0
screeninfo==0.8.1
seaborn==0.13.2
Send2Trash==1.8.3
setuptools==73.0.1
sh==2.0.7
shellingham==1.5.4
simple-websocket==1.1.0
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
sounddevice==0.5.1
soupsieve==2.7
SpeechRecognition==3.13.0
SQLAlchemy==2.0.38
sqlparse==0.5.1
stack-data==0.6.3
starlette==0.37.2
streamlit==1.41.1
sympy==1.14.0
tbb==2021.13.0
tenacity==9.0.0
tensorboard==2.17.0
tensorboard-data-server==0.7.2
tensorflow==2.17.0
tensorflow-intel==2.17.0
tensorrt==10.11.0.33
tensorrt_cu12==10.11.0.33
tensorrt_cu12_bindings==10.11.0.33
tensorrt_cu12_libs==10.11.0.33
termcolor==2.4.0
terminado==0.18.1
threadpoolctl==3.5.0
tinycss2==1.4.0
tk==0.1.0
tokenizers==0.21.1
toml==0.10.2
torch==2.7.1
torchaudio==2.7.1
torchvision==0.22.1
tornado==6.4.2
tqdm==4.67.1
tracker==0.1.1
traitlets==5.14.3
transformers==4.51.3
twilio==9.6.0
typer==0.12.3
types-python-dateutil==2.9.0.20250822
typing_extensions==4.12.2
tzdata==2024.1
ujson==5.10.0
ultralytics==8.3.154
ultralytics-thop==2.0.14
uri-template==1.3.0
uritemplate==4.1.1
urllib3==2.2.2
utils==1.0.2
uvicorn==0.30.2
virtualenv==20.26.3
watchdog==6.0.0
watchfiles==0.22.0
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
websockets==12.0
Werkzeug==3.0.3
wheel==0.44.0
widgetsnbextension==4.0.14
wrapt==1.16.0
wsproto==1.2.0
xyzservices==2024.6.0
yarl==1.20.0
zipp==3.21.0
zope.interface==6.4.post2
